feat: Setup call filters for testnet/mainnet
