+++ runtime/mainnet/src/lib.rs
@@ -21,6 +21,7 @@
 #[cfg(feature = "std")]
 include!(concat!(env!("OUT_DIR"), "/wasm_binary.rs"));
 
+mod filters;
 pub mod frontier_evm;
 pub mod impls;
 pub mod precompiles;
@@ -186,7 +187,7 @@ pub mod opaque {
 impl frame_system::Config for Runtime {
 	type AccountData = pallet_balances::AccountData<Balance>;
 	type AccountId = AccountId;
-	type BaseCallFilter = BaseFilter;
+	type BaseCallFilter = filters::MainnetCallFilter;
 	type BlockHashCount = BlockHashCount;
 	type BlockLength = BlockLength;
 	type Block = Block;
@@ -1024,39 +1025,6 @@ impl pallet_airdrop_claims::Config for Runtime {
 	type MaxVestingSchedules = MaxVestingSchedules;
 	type MoveClaimOrigin = frame_system::EnsureRoot<AccountId>;
 	type WeightInfo = TestWeightInfo;
-}
-
-pub struct BaseFilter;
-impl Contains<RuntimeCall> for BaseFilter {
-	fn contains(call: &RuntimeCall) -> bool {
-		let is_core_call = matches!(call, RuntimeCall::System(_) | RuntimeCall::Timestamp(_));
-		if is_core_call {
-			// always allow core call
-			return true
-		}
-
-		let is_paused =
-			pallet_transaction_pause::PausedTransactionFilter::<Runtime>::contains(call);
-		if is_paused {
-			// no paused call
-			return false
-		}
-
-		let democracy_related = matches!(
-			call,
-			// Filter democracy proposals creation
-			RuntimeCall::Democracy(_) |
-			// disallow council
-			RuntimeCall::Council(_)
-		);
-
-		if democracy_related {
-			// no democracy call
-			return false
-		}
-
-		true
-	}
 }
 
 // Create the runtime by composing the FRAME pallets that were previously configured.
