+++ runtime/testnet/src/lib.rs
@@ -21,6 +21,7 @@
 #[cfg(feature = "std")]
 include!(concat!(env!("OUT_DIR"), "/wasm_binary.rs"));
 
+mod filters;
 pub mod frontier_evm;
 pub mod impls;
 pub mod precompiles;
@@ -199,7 +200,7 @@ pub mod opaque {
 impl frame_system::Config for Runtime {
 	type AccountData = pallet_balances::AccountData<Balance>;
 	type AccountId = AccountId;
-	type BaseCallFilter = BaseFilter;
+	type BaseCallFilter = filters::TestnetCallFilter;
 	type BlockHashCount = BlockHashCount;
 	type BlockLength = BlockLength;
 	type Block = Block;
@@ -1139,39 +1140,6 @@ impl pallet_zksaas::Config for Runtime {
 	type UpdateOrigin = EnsureRootOrHalfCouncil;
 	type Verifier = (ArkworksVerifierGroth16Bn254, CircomVerifierGroth16Bn254);
 	type WeightInfo = ();
-}
-
-pub struct BaseFilter;
-impl Contains<RuntimeCall> for BaseFilter {
-	fn contains(call: &RuntimeCall) -> bool {
-		let is_core_call = matches!(call, RuntimeCall::System(_) | RuntimeCall::Timestamp(_));
-		if is_core_call {
-			// always allow core call
-			return true
-		}
-
-		let is_paused =
-			pallet_transaction_pause::PausedTransactionFilter::<Runtime>::contains(call);
-		if is_paused {
-			// no paused call
-			return false
-		}
-
-		let democracy_related = matches!(
-			call,
-			// Filter democracy proposals creation
-			RuntimeCall::Democracy(_) |
-			// disallow council
-			RuntimeCall::Council(_)
-		);
-
-		if democracy_related {
-			// no democracy call
-			return false
-		}
-
-		true
-	}
 }
 
 // Create the runtime by composing the FRAME pallets that were previously configured.
